minor modify

+++ script/parameter.py
@@ -11,14 +11,14 @@ def _set_fingerprint_params(conf: dict) -> None:
     SCAN_SPAN = float(conf["scan_span"])                      # length of span to scan [s]
     MARGIN_SPAN = float(conf["margin_span"])                  # length of span to prepare [s]
     SET_POINTS_POLICY = np.int8(conf["set_points_policy"])    # 1: ground truth of trajectory, 2: scan point file
-    XLIM = np.array(conf["xlim"], dtype=int)                  # default width limitation for map image
-    YLIM = np.array(conf["ylim"], dtype=int)                  # default height limitation for map image
+    XLIM = np.array(conf["xlim"], dtype=np.uint16)            # default width limitation for map image
+    YLIM = np.array(conf["ylim"], dtype=np.uint16)            # default height limitation for map image
 
 def _set_segment_params(conf: dict) -> None:
     global SEG_POLICY, MIN_COUNT
 
     SEG_POLICY = np.int8(conf["seg_policy"])                  # 1: median, 2: mode
-    MIN_COUNT = np.int64(conf["min_reception_count"])         # minimum count of signal reception to rely RSSI
+    MIN_COUNT = np.uint8(conf["min_reception_count"])         # minimum count of signal reception to rely RSSI
 
 def set_params(conf_file: Union[str, None] = None) -> dict:
     global ROOT_DIR
